requirements.txt
arrow==1.3.0
attrs==23.2.0
bio==1.7.1
biopython==1.83
biothings-client==0.3.1
boto3==1.34.141
botocore==1.34.141
bravado==11.0.3
bravado-core==6.1.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
DendroPy==5.0.1
ete3==3.1.3
filelock==3.14.0
fonttools==4.53.0
fqdn==1.5.1
fsspec==2024.6.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
gprofiler-official==1.0.0
idna==3.7
importlib_resources==6.4.0
isoduration==20.11.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.8
mygene==3.2.2
neptune==1.10.4
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.1
psutil==6.0.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.0
s3transfer==0.10.2
scipy==1.13.1
seaborn==0.13.2
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
swagger-spec-validator==3.0.4
sympy==1.12.1
torch==2.3.0
tqdm==4.66.4
triton==2.3.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.1
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
webcolors==24.6.0
websocket-client==1.8.0
